fix: Conserte o banco de dados e defina os triggers

+++ src/models/clientes.js
@@ -4,7 +4,7 @@ import sequelize from '../config/database.js';
 
 const Clientes = sequelize.define("Clientes", {
   clienteID: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.INTEGER(10),
     autoIncrement: true,
     primaryKey: true,
   },
